Fix typo in package name
