+++ tools/bamm-cli/src/main/java/io/openmanufacturing/sds/aspect/to/AspectToDotCommand.java
@@ -18,7 +18,7 @@ import java.io.IOException;
 import io.openmanufacturing.sds.AbstractCommand;
 import io.openmanufacturing.sds.aspect.AspectToCommand;
 import io.openmanufacturing.sds.aspectmodel.generator.diagram.AspectModelDiagramGenerator;
-import io.openmanufacturing.sds.exeption.CommandException;
+import io.openmanufacturing.sds.exception.CommandException;
 import picocli.CommandLine;
 
 @CommandLine.Command( name = AspectToDotCommand.COMMAND_NAME,
